Proposed changes: 1) changing tBTC token by adding (deprecated) to the name and tBTCv1 as symbol. 2) adding tBTCv2 with new contract number
Why this update?
tBTC v1 is the original version of tBTC that launched in September 2020 by Keep Network and has been winding down.

tBTC v2 launched . It's the successor of tBTC v1, which dramatically increases scalability through capital efficiency.

We want to minimize the chance for confusion for your users.

Adding 'deprecated' to the tBTC v1 token listings, will reduce the chance of someone attempting to acquire tBTC v2 to acquire tBTC v1

Adding tBTC v2 as a new listing, with the correct logo, will increase the chance that someone that attempts to acquire tBTC v2 will select the correct asset.

Still need to add the logo to the token-logos database as tbtc2.jpg